requirements.txt
requests
bs4
flask
flask_login
python-dotenv=1.0.1
firebase-admin
gunicorn
